refactor: move inference operation into goose-agent
