+++ crates/goose-agent/Cargo.toml
@@ -20,3 +20,6 @@ goose-provider-types = { version = "0.1.0-alpha.6", path = "../goose-provider-ty
 rmcp = { workspace = true }
 uuid = { workspace = true }
 serde_json = { workspace = true }
+futures = { workspace = true }
+tracing = { workspace = true }
+tracing-futures = { workspace = true }
