+++ crates/goose-agent/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod events;
+pub mod inference;
 pub mod machine;
 pub mod operation;
